Refactoring: Moved trace provider to internal package
