--- backend/tracing/propagator.go
@@ -1,8 +1,0 @@
-package tracing
-
-type PropagatorFormat string
-
-const (
-	PropagatorFormatJaeger PropagatorFormat = "jaeger"
-	PropagatorFormatW3C    PropagatorFormat = "w3c"
-)
